+++ Saltitantes/test/dominio/ClusterTest.java
@@ -1,5 +1,7 @@
-package model.entities;
+package dominio;
 
+import model.entities.Cluster;
+import model.entities.Duende;
 import model.entities.interfaces.EntityOnHorizon;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
